# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new client",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client successfully registered."
          },
          "400": {
            "description": "Bad request."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login a client",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Client successfully logged in."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Get current client profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Client profile retrieved successfully."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/projects": {
      "post": {
        "operationId": "ProjectsController_create",
        "summary": "Create a new project",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project successfully created."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProjectsController_findAll",
        "summary": "Get all projects (admin) or client's projects",
        "parameters": [
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Columns to select (comma-separated)",
            "schema": {
              "example": "id,name,description"
            }
          },
          {
            "name": "searchBy",
            "required": false,
            "in": "query",
            "description": "Columns to search in (comma-separated)",
            "schema": {
              "example": "name,description"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term",
            "schema": {
              "example": "project"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort by column and direction (e.g., name:ASC)",
            "schema": {
              "example": "id:ASC"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": 1
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects retrieved successfully."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "summary": "Get a project by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project retrieved successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "summary": "Update a project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProjectsController_remove",
        "summary": "Delete a project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/projects/{id}/matches/rebuild": {
      "post": {
        "operationId": "MatchingController_rebuildMatches",
        "summary": "Rebuild matches for a project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matches rebuilt successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "tags": [
          "matching"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/analytics/top-vendors": {
      "get": {
        "operationId": "AnalyticsController_getTopVendors",
        "summary": "Get top vendors by country",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Top vendors retrieved successfully."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Expansion Management API",
    "description": "API for managing client projects, vendor matching, and analytics",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}